Bugfix for referencing folder_from_settings variable outside its declaration scope.
Before, the folder_from_settings variable in save_file() was declared inside the if condition.
In my case, this was leading to a problem, for example if I call the screenshot-tool from the command line
using "screenshot-tool -s", the screenshot most of the times can't be saved. The reason for that is,
that when the folder_from_settings variable is declared inside the if condition and assigned to the folder_dir
variable (which is used outside the if condition later), the memory for the folder_from_settings variable
might be already released. By playing its declaration outside the if condition it works now for me.