patch: patch screenshot crashing when height exceeds width
Closes #316.

## Description

This addresses the issue discussed in issue #316. To summarize what is in that issue, when taking a screenshot and the height is 1/4x or more greater than the width (approx), it will segfault.

Two solutions were found, but the simplest was in this PR, which is setting a default window height. The other was to use a deprecated method (which is shown in the issue).